Reorganizing the evaluation module (#34)
