+++ src/midst_toolkit/evaluation/privacy/scripts/midst_dcr_score_eval.py
@@ -6,7 +6,7 @@ import pandas as pd
 
 from midst_toolkit.common.logger import log
 from midst_toolkit.data_processing.midst_data_processing import load_midst_data_with_test
-from midst_toolkit.evaluation.generation.distance_closest_record import (
+from midst_toolkit.evaluation.privacy.distance_closest_record import (
     DistanceToClosestRecordScore,
     preprocess,
 )
