+++ tests/unit/evaluation/privacy/test_median_dcr.py
@@ -5,12 +5,12 @@ import pytest
 import torch
 
 from midst_toolkit.data_processing.midst_data_processing import load_midst_data
-from midst_toolkit.evaluation.generation.distance_closest_record import (
+from midst_toolkit.evaluation.privacy.distance_closest_record import (
     MedianDistanceToClosestRecordScore,
     minimum_distances,
     preprocess,
 )
-from midst_toolkit.evaluation.generation.distance_utils import NormType
+from midst_toolkit.evaluation.privacy.distance_utils import NormType
 
 
 SYNTHETIC_DATA = torch.Tensor([[1.0, 2.0, 1.0], [1.0, 2.0, 3.0], [3.4, 1.0, 0.3]])
