+++ tests/unit/evaluation/quality/test_confidence_interval_overlap.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import pytest
 
-from midst_toolkit.evaluation.generation.confidence_interval_overlap import (
+from midst_toolkit.evaluation.quality.confidence_interval_overlap import (
     ConfidenceLevel,
     MeanConfidenceInternalOverlap,
 )
