+++ tests/unit/evaluation/quality/test_utils.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 import pytest
 
-from midst_toolkit.evaluation.generation.utils import (
+from midst_toolkit.evaluation.utils import (
     create_quality_metrics_directory,
     dump_metrics_dict,
 )
